requirements.txt
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
click==8.1.7
Flask==3.0.3
googletrans==4.0.0rc1
google-cloud-translate==3.9.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy
pandas
protobuf>=5.26.1,<6.0dev
python-dateutil==2.9.0.post0
pytz==2024.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
tzdata==2024.1
Werkzeug==3.0.3
gunicorn==20.1.0